feat(jans-auth-server): rely on end_session feature flag value when return discovery response #5255
