+++ jans-auth-server/server/src/test/resources/testng.xml
@@ -23,6 +23,7 @@
             <class name="io.jans.as.server.service.GrantServiceTest" />
             <class name="io.jans.as.server.service.LocalResponseCacheTest" />
             <class name="io.jans.as.server.service.SessionIdServiceTest" />
+            <class name="io.jans.as.server.service.DiscoveryServiceTest" />
 
             <class name="io.jans.as.server.auth.DpopServiceTest" />
             <class name="io.jans.as.server.token.ws.rs.TokenExchangeServiceTest" />
